feat: add cli and update readme
